breaking: changes to project structure, various fixes to routes, added frontend and start scripts, modified environment variables...

+++ src/controllers/ProductController.ts
@@ -17,9 +17,11 @@ export class ProductController {
     @Logger
     static async createProduct(req: Request, res: Response) {
         try {
+            if (req.statusCode === 400) return res.status(400);
             const nuevoProducto = new Product(req.body);
             await nuevoProducto.save();
             return res.status(201).json(nuevoProducto);
+
         } catch (error) {
         }
     }
@@ -27,6 +29,7 @@ export class ProductController {
     @Logger
     static async updateProduct(req: Request, res: Response) {
         try {
+            if (req.statusCode === 400) return res.status(400);
             const producto = await Product.findByIdAndUpdate(req.params.id, req.body, { new: true });
             if (!producto) return res.status(404).json({ message: "Producto no encontrado" });
             return res.json(producto);
@@ -36,6 +39,7 @@ export class ProductController {
 
     static async deleteProduct(req: Request, res: Response) {
         try {
+            if (req.statusCode === 400) return res.status(400);
             const producto = await Product.findByIdAndDelete(req.params.id);
             if (!producto) return res.status(404).json({ message: "Producto no encontrado" });
             return res.json({ message: "Producto eliminado" });
